Remove Discord default title format
It was an okay idea but it's the only title that isn't a custom regex, and it would be hard to maintain. Also, Discord does it differently per client. Also, users would have to wait 2 seconds if they included an embed. I just don't think this feature is worth it atm.